breadcrumb patch and utils file
